+++ src/jap/JAPNewView.java
@@ -2919,7 +2919,7 @@ final public class JAPNewView extends AbstractJAPMainView implements IJAPMainVie
 				{
 					m_labelDelay.setText(JAPMessages.getString("statusUnknown"));
 				}
-				else if(value == Long.MAX_VALUE)
+				else if(value == Integer.MAX_VALUE)
 				{
 					m_labelDelay.setText("> " + 
 							PerformanceEntry.BOUNDARIES[PerformanceEntry.DELAY][
